move getCanisterActor into separte file to avoid jest
